extract the parts annotation compute into a shared lib and add a resolvedParts copy alongside it
